+++ src/main/java/se/sundsvall/casedata/integration/db/NotificationRepository.java
@@ -1,9 +1,12 @@
 package se.sundsvall.casedata.integration.db;
 
 import java.time.OffsetDateTime;
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
+import jakarta.validation.constraints.NotNull;
+import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
@@ -12,11 +15,41 @@ import se.sundsvall.casedata.integration.db.model.NotificationEntity;
 @CircuitBreaker(name = "notificationRepository")
 public interface NotificationRepository extends JpaRepository<NotificationEntity, String> {
 
-	Optional<NotificationEntity> findByIdAndNamespaceAndMunicipalityId(String id, String namespace, String municipalityId);
+	boolean existsByIdAndNamespaceAndMunicipalityIdAndErrandId(
+		final String id,
+		final String namespace,
+		final String municipalityId,
+		final Long errandId);
 
-	List<NotificationEntity> findAllByNamespaceAndMunicipalityIdAndOwnerId(String namespace, String municipalityId, String ownerId);
+	Optional<NotificationEntity> findByIdAndNamespaceAndMunicipalityId(
+		String id,
+		String namespace,
+		String municipalityId);
 
-	Optional<NotificationEntity> findByNamespaceAndMunicipalityIdAndOwnerIdAndAcknowledgedAndErrandIdAndType(String namespace, String municipalityId, String ownerId, boolean acknowledged, long errandId, String type);
+	List<NotificationEntity> findAllByNamespaceAndMunicipalityIdAndOwnerId(
+		String namespace,
+		String municipalityId,
+		String ownerId);
+
+	Optional<NotificationEntity> findByNamespaceAndMunicipalityIdAndOwnerIdAndAcknowledgedAndErrandIdAndType(
+		String namespace,
+		String municipalityId,
+		String ownerId,
+		boolean acknowledged,
+		long errandId,
+		String type);
 
 	List<NotificationEntity> findByExpiresBefore(final OffsetDateTime expires);
+
+	List<NotificationEntity> findAllByNamespaceAndMunicipalityIdAndErrandId(
+		String namespace,
+		String municipalityId,
+		Long errandId,
+		Sort sort);
+
+	Optional<NotificationEntity> findByIdAndNamespaceAndMunicipalityIdAndErrandId(
+		String notificationId,
+		String namespace,
+		String municipalityId,
+		Long errandId);
 }