+++ src/main/java/se/sundsvall/casedata/service/NotificationService.java
@@ -9,6 +9,7 @@ import static se.sundsvall.casedata.service.util.mappers.PatchMapper.patchNotifi
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.zalando.problem.Problem;
 
@@ -23,8 +24,8 @@ import se.sundsvall.casedata.service.util.mappers.EntityMapper;
 @Service
 public class NotificationService {
 
-	private static final String NOTIFICATION_ENTITY_NOT_FOUND = "Notification with id '%s' not found in namespace '%s' for municipality with id '%s'";
-	private static final String ERRAND_ENTITY_NOT_FOUND = "Errand with id '%'s not found in namespace '%s' for municipality with id '%s'";
+	private static final String NOTIFICATION_ENTITY_NOT_FOUND = "Notification with id:'%s' not found in namespace:'%s' for municipality with id:'%s' and errand with id:'%s'";
+	private static final String ERRAND_ENTITY_NOT_FOUND = "Errand with id:'%s' not found in namespace:'%s' for municipality with id:'%s'";
 
 	private final IncomingRequestFilter incomingRequestFilter;
 	private final NotificationRepository notificationRepository;
@@ -39,7 +40,7 @@ public class NotificationService {
 		ACCESS,
 		EXECUTE,
 		CANCEL,
-		DROP;
+		DROP
 	}
 
 	public NotificationService(
@@ -54,19 +55,28 @@ public class NotificationService {
 		this.employeeService = employeeService;
 	}
 
-	public List<Notification> getNotifications(String municipalityId, String namespace, String ownerId) {
-		return notificationRepository.findAllByNamespaceAndMunicipalityIdAndOwnerId(namespace, municipalityId, ownerId).stream()
+	public List<Notification> getNotificationsByOwnerId(final String municipalityId, final String namespace, final String ownerId) {
+		return notificationRepository.findAllByNamespaceAndMunicipalityIdAndOwnerId(namespace, municipalityId, ownerId)
+			.stream()
 			.map(EntityMapper::toNotification)
 			.toList();
 	}
 
-	public Notification getNotification(String municipalityId, String namespace, String notificationId) {
-		return notificationRepository.findByIdAndNamespaceAndMunicipalityId(notificationId, namespace, municipalityId)
+	public List<Notification> getNotificationsByErrandId(final String municipalityId, final String namespace, final Long errandId, final Sort sort) {
+		return notificationRepository.findAllByNamespaceAndMunicipalityIdAndErrandId(namespace, municipalityId, errandId, sort)
+			.stream()
 			.map(EntityMapper::toNotification)
-			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, NOTIFICATION_ENTITY_NOT_FOUND.formatted(notificationId, namespace, municipalityId)));
+			.toList();
 	}
 
-	public Notification createNotification(String municipalityId, String namespace, Notification notification) {
+
+	public Notification getNotification(final String municipalityId, final String namespace,final Long errandId, final String notificationId) {
+		return notificationRepository.findByIdAndNamespaceAndMunicipalityIdAndErrandId(notificationId, namespace, municipalityId, errandId)
+			.map(EntityMapper::toNotification)
+			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, NOTIFICATION_ENTITY_NOT_FOUND.formatted(notificationId, namespace, municipalityId,errandId)));
+	}
+
+	public Notification createNotification(final String municipalityId, final String namespace, final Notification notification) {
 		if ((notification.getOwnerId() == null) || isExecutingUserTheOwner(notification.getOwnerId()) || notificationExists(municipalityId, namespace, notification)) {
 			return null;
 		}
@@ -81,20 +91,21 @@ public class NotificationService {
 
 	}
 
-	public void updateNotifications(String municipalityId, String namespace, List<PatchNotification> notifications) {
+	public void updateNotifications(final String municipalityId, final String namespace, final List<PatchNotification> notifications) {
 		notifications.forEach(notification -> updateNotification(municipalityId, namespace, notification.getId(), notification));
 	}
 
-	public void deleteNotification(String municipalityId, String namespace, String notificationId) {
-		final var entity = notificationRepository.findByIdAndNamespaceAndMunicipalityId(notificationId, namespace, municipalityId)
-			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, NOTIFICATION_ENTITY_NOT_FOUND.formatted(notificationId, namespace, municipalityId)));
+	public void deleteNotification(final String municipalityId, final String namespace, final Long errandId, final String notificationId) {
 
-		notificationRepository.delete(entity);
+		if (!notificationRepository.existsByIdAndNamespaceAndMunicipalityIdAndErrandId(notificationId, namespace, municipalityId, errandId)) {
+			throw Problem.valueOf(NOT_FOUND, String.format(NOTIFICATION_ENTITY_NOT_FOUND, notificationId, namespace, municipalityId, errandId));
+		}
+		notificationRepository.deleteById(notificationId);
 	}
 
 	private void updateNotification(final String municipalityId, final String namespace, final String notificationId, final PatchNotification notification) {
-		final var entity = notificationRepository.findByIdAndNamespaceAndMunicipalityId(notificationId, namespace, municipalityId)
-			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, NOTIFICATION_ENTITY_NOT_FOUND.formatted(notificationId, namespace, municipalityId)));
+		final var entity = notificationRepository.findByIdAndNamespaceAndMunicipalityIdAndErrandId(notificationId, namespace, municipalityId, notification.getErrandId())
+			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, String.format(NOTIFICATION_ENTITY_NOT_FOUND, notificationId, namespace, municipalityId, notification.getErrandId())));
 
 		final var owner = getPortalPersonData(notification.getOwnerId());
 
@@ -117,7 +128,7 @@ public class NotificationService {
 		return equalsIgnoreCase(ownerId, incomingRequestFilter.getAdUser());
 	}
 
-	private PortalPersonData getPortalPersonData(String adAccountId) {
+	private PortalPersonData getPortalPersonData(final String adAccountId) {
 		return Optional.ofNullable(adAccountId)
 			.map(employeeService::getEmployeeByLoginName)
 			.orElse(null);