+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## API-endpoints
 
+### Attachments:
+#### New endpoints:
+      - [GET] /{municipalityId}/{namespace}/errand/{errandNumber}/attachments
+#### Removed endpoints:
+      - [GET] /{municipalityId}/{namespace}/attachments/errand/{errandNumber}
 ### Appeals:
 
 #### New endpoints:
@@ -14,6 +19,17 @@
       - [GET] /{municipalityId}/{namespace}/errands/{errandId}/appeals/{appealId}
       - [PATCH] /{municipalityId}/{namespace}/errands/{errandId}/appeals/{appealId}
       - [PUT] /{municipalityId}/{namespace}/errands/{errandId}/appeals/{appealId}
+
+### Notifications:
+#### New endpoints:
+      - [GET] /{municipalityId}/{namespace}/errands/{errandId}/notifications
+      - [POST] /{municipalityId}/{namespace}/errands/{errandId}/notifications
+      - [DELETE] /{municipalityId}/{namespace}/errands/{errandId}/notifications/{notificationId}
+      - [GET] /{municipalityId}/{namespace}/errands/{errandId}/notifications/{notificationId}
+#### Removed endpoints:
+      - [POST] /{municipalityId}/{namespace}/notifications
+      - [DELETE] /{municipalityId}/{namespace}/notifications/{notificationId}
+      - [GET] /{municipalityId}/{namespace}/notifications/{notificationId}
 
 ## API-Model updates
 
@@ -35,8 +51,20 @@
 		- `ANMALAN_HALSOSKYDDSVERKSAMHET`
 
 - **Errand**
-	- **Removed Fields:**
-		- appeals: ``List<Appeal>``
+	- **Added Fields:**
+        - relatesTo: `List<RelatedErrand>`
+    - **Removed Fields:**
+		- appeals: `List<Appeal>`
+      
+- **PatchNotification**
+	- **Added Fields:**
+		- errandId: `Long`
+
+- **RelatedErrand** *(Added)*
+    - **Fields:**
+        - errandId: `Long`
+        - errandNumber: `String`
+        - relationReason: `String`
 
 - **Appeal** *(Removed)*
 	- **Fields:**
@@ -78,7 +106,6 @@
       - [PATCH] /{municipalityId}/{namespace}/errands/{errandId}/appeals/{appealId}
       - [PUT] /{municipalityId}/{namespace}/errands/{errandId}/appeals/{appealId}
 
->>>>>>> 1bc0676 (UF-10650: Removed appeal object, entity and supporting classes)
 #### Removed endpoints:
 
       - [GET] /{municipalityId}/appeals/{appealId}